auth-service/requirements.txt
fastapi
uvicorn
python-jose
prometheus-client
PyJWT